add CssFile, css-imports and Frame-class
differentiate between CssFile and JsFile when processing imports.

JsFile strips its css-imports saves all the css-imports it encounters during the typescript transform and saves the css-files to the cssImports-signal. It then strips them from the code.

CssFile creates a module-url with a Blob. Thanks @modderme for the code to inject the css.

These imports can then be used in onReady to injects css into a frame's window.

Frame-class with injectFile-method.