add custom-element wrappers for the headless components
support for
- <repl-frame/>
- <repl-devtools/>
- <repl-shiki-editor/>
- <repl-monaco-editor/>

Instead of a root <Repl/>-component to set the runtime, we provide
now setRuntime-function that sets a global runtime.

Similarly <repl-monaco-editor/> is set with the function setMonaco.

The solid components Frame, DevTools, ShikiEditor and MonacoEditor now
have a compound component .Standalone that needs to be passed a runtime-prop.
MonacoEditor.Standalone also needs to be passed a Monaco-instance.
This way the custom elements can be tiny wrappers around the compound
components.