remove FrameRegistry from runtime
the runtime had a frame-registry from which you could query the frames that were in the markup. this worked from the assumption that everything was nested inside a `<Repl/>` component. this assumption is not correct anymore.

i was never a fan of using ids to get the frame from the runtime, and there is currently better api to get access to a specific frame: onFrame of the `<repl-frame/>` element. this allows us to colocalise the frame actions with the frame itself.

repl-frame can now be used completely independent from the runtime.